+++ src/mappings/Season2019/Attributes/Track/index.js
@@ -1,0 +1,15 @@
+const Tracks = require('./Tracks');
+
+const ByName = {};
+const ById = {};
+
+for (const track of Tracks) {
+    ByName[track.track] = track;
+    ById[track.trackId] = track;
+}
+
+module.exports = {
+    All: Tracks,
+    ByName,
+    ById,
+};