+++ src/mappings/Season2019/TokenTypes/index.js
@@ -1,0 +1,13 @@
+const Car = require('./Car');
+const Driver = require('./Driver');
+const Part = require('./Part');
+const Gear = require('./Gear');
+const Tyres = require('./Tyres');
+
+module.exports = {
+    Car,
+    Driver,
+    Part,
+    Gear,
+    Tyres,
+}