+++ src/mappings/Season2019/index.js
@@ -1,25 +1,13 @@
 const Season = require('./Season');
 const Collection = require('./Collection');
 const SubType = require('./SubType');
-const Car = require('./Car');
-const Driver = require('./Driver');
-const Part = require('./Part');
-const Gear = require('./Gear');
-const Tyres = require('./Tyres');
-const Team = require('./Team');
-const GrandPrix = require('./GrandPrix');
-const Model = require('./Model');
+const Attributes = require('./Attributes');
+const TokenTypes = require('./TokenTypes');
 
 module.exports = {
     Season,
     Collection,
     SubType,
-    Car,
-    Driver,
-    Part,
-    Gear,
-    Tyres,
-    Team,
-    GrandPrix,
-    Model,
+    Attributes,
+    TokenTypes,
 };