+++ src/mappings/Season2020/Attributes/index.js
@@ -1,0 +1,9 @@
+const Model = require('./Model');
+const Team = require('./Team');
+const Track = require('./Track');
+
+module.exports = {
+    Model,
+    Team,
+    Track,
+}